# [pt][quant] Add vector path to copy kernel for quantized data types