@dskhudia 2020-04-09
> Can you also provide a simple benchmark script as part of the diff comments and show the perf improvement? It could just be copying a quantized tensor within a loop.

Updated summary and added numbers.